requirements.txt
# 核心依赖
pydantic==1.10.13
python-dotenv==1.0.0

# API 框架
fastapi==0.104.1
uvicorn==0.24.0

# 模型和 AI 相关
torch==2.1.0
transformers==4.35.0
tokenizers>=0.14,<0.15
spacy==3.5.3
presidio-analyzer==2.2.33
presidio-anonymizer==2.2.33

# HTTP 客户端
requests==2.31.0
aiohttp==3.9.1

# 工具库
numpy==1.24.3
python-multipart==0.0.6

# 其他依赖
streamlit==1.22.0
langdetect==1.0.9
python-jose[cryptography]==3.3.0
bcrypt==4.1.2
sentencepiece==0.1.99
loguru==0.6.0 